pyproject.toml
[tool.poetry]
name = "ai-gym-bro"
version = "0.1.0"
description = "Telegram bot with OpenAI integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
run-bot = "ai_gym_bro.main:main"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "20.7"
openai = "1.12.0"
python-dotenv = "1.0.0"
ruff = "^0.11.5"
loguru = "^0.7.3"
langfuse = "^2.60.5"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"
pytest-asyncio = "0.23.5"
mypy = "1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"
lint.select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF", "I"]
lint.ignore = []

# Enable pyupgrade and isort rules.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 